+++ public/assets/js/asignar-turnos.js
@@ -103,7 +103,7 @@ document.addEventListener('DOMContentLoaded', function () {
                         <td>${turno.areas_count}</td>
                         <td>
                             <img src="assets/icons/editar.png" alt="Editar" class="icono-editar editarTurnoBtn" data-id="${turno.id}">
-                            <img src="assets/icons/eliminar.png" alt="Eliminar" class="icono-eliminar" data-id="${turno.id}">
+                            <img src="assets/icons/eliminar.png" alt="Eliminar" class="icono-eliminar" data-id="${turno.id}" data-tipo="asignar-turno">
                         </td>
                     `;
                     turnosTableBody.appendChild(row);
@@ -299,7 +299,7 @@ document.addEventListener('DOMContentLoaded', function () {
         if (event.target.classList.contains('editarTurnoBtn')) {
             turnoId = event.target.getAttribute('data-id');
             document.getElementById('asignarTurnoModal').setAttribute('data-turno-id', turnoId); // Almacenar el turnoId en el modal
-            asignarTurnoModalLabel.textContent = "Editar Turno Asignado #"+turnoId;
+            asignarTurnoModalLabel.textContent = "Editar Turno Asignado #" + turnoId;
             console.log('Id: ', turnoId)
             editMode = true;
 
@@ -329,30 +329,33 @@ document.addEventListener('DOMContentLoaded', function () {
 
         if (event.target.classList.contains('icono-eliminar')) {
             const turnoId = event.target.getAttribute('data-id');
+            const tipo = event.target.getAttribute("data-tipo");
             if (!turnoId) {
                 console.error("Error: No se encontró el ID del turno en el botón.");
                 return;
             }
-
-            if (confirm("¿Está seguro de que desea eliminar este turno asignado?")) {
-                fetch(`asignar-turnos/eliminar/${turnoId}`, {
-                    method: "DELETE",
-                    headers: {
-                        'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content')
-                    }
-                })
-                .then(response => {
-                    if (!response.ok) {
-                        throw new Error("Error al eliminar el turno asignado.");
-                    }
-                    return response.json();
-                })
-                .then(data => {
-                    alert("Turno asignado eliminado con éxito.");
-                    consultarTurnosAsignados(); // Reload the table
-                })
-                .catch(error => console.error("Error al eliminar el turno asignado:", error));
+            if (tipo === "asignar-turno") {
+                if (confirm("¿Está seguro de que desea eliminar este turno asignado?")) {
+                    fetch(`asignar-turnos/eliminar/${turnoId}`, {
+                        method: "DELETE",
+                        headers: {
+                            'X-CSRF-TOKEN': document.querySelector('meta[name="csrf-token"]').getAttribute('content')
+                        }
+                    })
+                        .then(response => {
+                            if (!response.ok) {
+                                throw new Error("Error al eliminar el turno asignado.");
+                            }
+                            return response.json();
+                        })
+                        .then(data => {
+                            alert("Turno asignado eliminado con éxito.");
+                            consultarTurnosAsignados(); // Reload the table
+                        })
+                        .catch(error => console.error("Error al eliminar el turno asignado:", error));
+                }
             }
+
         }
     });
